model-service/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
python-dotenv